SSCS-8977 Remove edited bundle callback handler
