+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/createbundle/CreateBundleAboutToSubmitHandler.java
@@ -88,9 +88,9 @@ public class CreateBundleAboutToSubmitHandler implements PreSubmitCallbackHandle
 
         moveDocsToDwpCollectionIfOldPattern(sscsCaseData);
 
+        PreSubmitCallbackResponse<SscsCaseData> response = new PreSubmitCallbackResponse<>(callback.getCaseDetails().getCaseData());
+
         if (checkMandatoryFilesMissing(sscsCaseData)) {
-            PreSubmitCallbackResponse<SscsCaseData> response;
-            response = new PreSubmitCallbackResponse<>(callback.getCaseDetails().getCaseData());
             response.addError("The bundle cannot be created as mandatory DWP documents are missing");
             return response;
         } else {
@@ -120,14 +120,19 @@ public class CreateBundleAboutToSubmitHandler implements PreSubmitCallbackHandle
             bundleAudioVideoPdfService.createAudioVideoPdf(sscsCaseData);
 
             if (multiBundleFeature) {
-                setMultiBundleConfig(sscsCaseData);
+                setMultiBundleConfig(sscsCaseData, response);
             } else {
                 setBundleConfig(sscsCaseData);
             }
 
-            log.info("Setting the bundleConfiguration on the case {} for case id {}", sscsCaseData.getBundleConfiguration(), callback.getCaseDetails().getId());
+            if (response.getErrors() != null && !response.getErrors().isEmpty()) {
+                log.info("Error found in bundle creation process for case id {}", callback.getCaseDetails().getId());
+                return response;
+            } else {
+                log.info("Setting the bundleConfiguration on the case {} for case id {}", sscsCaseData.getBundleConfiguration(), callback.getCaseDetails().getId());
 
-            return serviceRequestExecutor.post(callback, bundleUrl + CREATE_BUNDLE_ENDPOINT);
+                return serviceRequestExecutor.post(callback, bundleUrl + CREATE_BUNDLE_ENDPOINT);
+            }
         }
     }
 
@@ -150,13 +155,18 @@ public class CreateBundleAboutToSubmitHandler implements PreSubmitCallbackHandle
         }
     }
 
-    private void setMultiBundleConfig(SscsCaseData sscsCaseData) {
+    private void setMultiBundleConfig(SscsCaseData sscsCaseData, PreSubmitCallbackResponse<SscsCaseData> response) {
         List<MultiBundleConfig> configs = new ArrayList<>();
 
         if (sscsCaseData.getDwpDocuments().stream().filter(f -> (f.getValue().getDocumentType().equals(DwpDocumentType.DWP_RESPONSE.getValue())
                 && f.getValue().getEditedDocumentLink() != null)
                 || f.getValue().getDocumentType().equals(DwpDocumentType.DWP_EVIDENCE_BUNDLE.getValue())
                 && f.getValue().getEditedDocumentLink() != null).count() > 0) {
+
+            if (checkPhmeStatusIsNotGranted(sscsCaseData)) {
+                response.addError("The edited bundle cannot be created as PHME status has not been granted");
+                return;
+            }
 
             if (sscsCaseData.isLanguagePreferenceWelsh()) {
                 configs.add(MultiBundleConfig.builder().value(bundleWelshEditedConfig).build());
@@ -207,5 +217,8 @@ public class CreateBundleAboutToSubmitHandler implements PreSubmitCallbackHandle
         }
     }
 
+    protected boolean checkPhmeStatusIsNotGranted(SscsCaseData sscsCaseData) {
+        return sscsCaseData.getPhmeGranted() == null || sscsCaseData.getPhmeGranted().getValue().equals("No");
+    }
 
 }
