+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/createbundle/CreateBundleAboutToSubmitHandlerTest.java
@@ -196,6 +196,8 @@ public class CreateBundleAboutToSubmitHandlerTest {
                 .findFirst()
                 .orElse("");
         assertEquals("The bundle cannot be created as mandatory DWP documents are missing", error);
+        verifyNoInteractions(serviceRequestExecutor);
+
     }
 
     @Test
@@ -209,6 +211,7 @@ public class CreateBundleAboutToSubmitHandlerTest {
                 .findFirst()
                 .orElse("");
         assertEquals("The bundle cannot be created as mandatory DWP documents are missing", error);
+        verifyNoInteractions(serviceRequestExecutor);
     }
 
     @Test
@@ -224,6 +227,7 @@ public class CreateBundleAboutToSubmitHandlerTest {
                 .findFirst()
                 .orElse("");
         assertEquals("The bundle cannot be created as mandatory DWP documents are missing", error);
+        verifyNoInteractions(serviceRequestExecutor);
     }
 
     @Test
@@ -237,6 +241,7 @@ public class CreateBundleAboutToSubmitHandlerTest {
                 .findFirst()
                 .orElse("");
         assertEquals("The bundle cannot be created as mandatory DWP documents are missing", error);
+        verifyNoInteractions(serviceRequestExecutor);
     }
 
     @Test
@@ -263,6 +268,8 @@ public class CreateBundleAboutToSubmitHandlerTest {
         dwpDocuments.add(DwpDocument.builder().value(DwpDocumentDetails.builder().documentType(DWP_RESPONSE.getValue()).documentLink(DocumentLink.builder().documentFilename("Testing").build()).editedDocumentLink(DocumentLink.builder().documentFilename("Testing").build()).build()).build());
         callback.getCaseDetails().getCaseData().setDwpDocuments(dwpDocuments);
 
+        callback.getCaseDetails().getCaseData().setDwpPhme("Yes");
+        callback.getCaseDetails().getCaseData().setPhmeGranted(YesNo.YES);
         callback.getCaseDetails().getCaseData().setLanguagePreferenceWelsh("No");
         PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
 
@@ -281,6 +288,8 @@ public class CreateBundleAboutToSubmitHandlerTest {
         dwpDocuments.add(DwpDocument.builder().value(DwpDocumentDetails.builder().documentType(DWP_RESPONSE.getValue()).documentLink(DocumentLink.builder().documentFilename("Testing").build()).editedDocumentLink(DocumentLink.builder().documentFilename("Testing").build()).build()).build());
         callback.getCaseDetails().getCaseData().setDwpDocuments(dwpDocuments);
 
+        callback.getCaseDetails().getCaseData().setDwpPhme("Yes");
+        callback.getCaseDetails().getCaseData().setPhmeGranted(YesNo.YES);
         callback.getCaseDetails().getCaseData().setLanguagePreferenceWelsh("Yes");
         PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
 
@@ -288,6 +297,28 @@ public class CreateBundleAboutToSubmitHandlerTest {
         assertEquals(2, response.getData().getMultiBundleConfiguration().size());
         assertEquals("bundleWelshEditedConfig", response.getData().getMultiBundleConfiguration().get(0).getValue());
         assertEquals("bundleWelshUnEditedConfig", response.getData().getMultiBundleConfiguration().get(1).getValue());
+    }
+
+    @Test
+    public void givenCaseWithEditedDwpDocsAndPhmeNotGrantedAndMultiBundleFeatureSwitchedOn_thenReturnErrorMessageAndDoNotSendRequestToBundleService() {
+        ReflectionTestUtils.setField(handler, "multiBundleFeature", true);
+
+        List<DwpDocument> dwpDocuments = new ArrayList<>();
+        dwpDocuments.add(DwpDocument.builder().value(DwpDocumentDetails.builder().documentType(DWP_EVIDENCE_BUNDLE.getValue()).documentLink(DocumentLink.builder().documentFilename("Testing").build()).editedDocumentLink(DocumentLink.builder().build()).build()).build());
+        dwpDocuments.add(DwpDocument.builder().value(DwpDocumentDetails.builder().documentType(DWP_RESPONSE.getValue()).documentLink(DocumentLink.builder().documentFilename("Testing").build()).editedDocumentLink(DocumentLink.builder().documentFilename("Testing").build()).build()).build());
+        callback.getCaseDetails().getCaseData().setDwpDocuments(dwpDocuments);
+
+        callback.getCaseDetails().getCaseData().setLanguagePreferenceWelsh("No");
+        callback.getCaseDetails().getCaseData().setDwpPhme("Yes");
+        callback.getCaseDetails().getCaseData().setPhmeGranted(YesNo.NO);
+
+        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
+
+        String error = response.getErrors().stream()
+                .findFirst()
+                .orElse("");
+        assertEquals("The edited bundle cannot be created as PHME status has not been granted", error);
+        verifyNoInteractions(serviceRequestExecutor);
     }
 
     @Test
